requirements.txt
flask
flask_sqlalchemy
flask_migrate
flask_jwt_extended
python-dotenv
pillow